+++ test/Unit/Ads/OpenMeasurement/OpenMeasurementTest.ts
@@ -1,6 +1,6 @@
 import * as sinon from 'sinon';
 import { assert } from 'chai';
-import { OpenMeasurement } from 'Ads/Views/OpenMeasurement';
+import { OpenMeasurement } from 'Ads/Views/OpenMeasurement/OpenMeasurement';
 import { Platform } from 'Core/Constants/Platform';
 import { ICoreApi } from 'Core/ICore';
 import { TestFixtures } from 'TestHelpers/TestFixtures';
@@ -11,13 +11,12 @@ import { Placement } from 'Ads/Models/Placement';
 import { DeviceInfo } from 'Core/Models/DeviceInfo';
 import { VastCampaign } from 'VAST/Models/VastCampaign';
 import { RequestManager } from 'Core/Managers/RequestManager';
-import { IAdView, ObstructionReasons, IRectangle } from 'Ads/Views/OMIDEventBridge';
-import OMID3p from 'html/omid/omid3p.html';
 import { VastAdVerification } from 'VAST/Models/VastAdVerification';
 import { VastVerificationResource } from 'VAST/Models/VastVerificationResource';
+import OMID3p from 'html/omid/omid3p.html';
 
 [Platform.ANDROID, Platform.IOS].forEach(platform => {
-    describe(`${platform} OpenMeasurement`, () => {
+    describe(`${platform} OpenMeasurementTest`, () => {
         const sandbox = sinon.createSandbox();
         let om: OpenMeasurement;
         let backend: Backend;
@@ -45,10 +44,10 @@ import { VastVerificationResource } from 'VAST/Models/VastVerificationResource';
 
             request = sinon.createStubInstance(RequestManager);
             if (verifications) {
-                return new OpenMeasurement(platform, core, clientInformation, campaign, placement, deviceInfo, request, verifications);
+                return new OpenMeasurement(platform, core, clientInformation, campaign, placement, deviceInfo, request, 'test', verifications[0]);
             } else {
-                const verification = campaign.getVast().getAdVerifications();
-                return new OpenMeasurement(platform, core, clientInformation, campaign, placement, deviceInfo, request, verification);
+                const verification = campaign.getVast().getAdVerifications()[0];
+                return new OpenMeasurement(platform, core, clientInformation, campaign, placement, deviceInfo, request, 'test', verification);
             }
         };
 
@@ -65,7 +64,7 @@ import { VastVerificationResource } from 'VAST/Models/VastVerificationResource';
 
                 it('should populate the omid-iframe with omid3p container code', () => {
                     om.render();
-                    assert.equal((<HTMLIFrameElement>om.container().querySelector('#omid-iframe')).srcdoc, OMID3p.replace('{{ DEFAULT_KEY_ }}', 'default_key'));
+                    assert.equal((<HTMLIFrameElement>om.container().querySelector('#omid-iframe' + om.getOMAdSessionId())).srcdoc, OMID3p.replace('{{ DEFAULT_KEY_ }}', 'default_key'));
                 });
 
                 it('should not call the remove child function if om does not exist in dom', () => {
@@ -101,10 +100,11 @@ import { VastVerificationResource } from 'VAST/Models/VastVerificationResource';
                         om.removeMessageListener();
                     });
 
-                    it('should populate script to dom with multiple passed resources', () => {
+                    // now we use iframes for multiple scripts
+                    it('should populate script to dom with one passed resource', () => {
                         // need a more reliable way to check the dom
-                        sinon.assert.calledTwice(<sinon.SinonStub>om.injectAsString);
-                        sinon.assert.calledTwice(<sinon.SinonStub>om.populateVendorKey);
+                        sinon.assert.calledOnce(<sinon.SinonStub>om.injectAsString);
+                        sinon.assert.calledOnce(<sinon.SinonStub>om.populateVendorKey);
                     });
                 });
 
@@ -159,28 +159,43 @@ import { VastVerificationResource } from 'VAST/Models/VastVerificationResource';
                 describe('onEventProcessed', () => {
                     context('sessionStart', () => {
                         beforeEach(() => {
+                            om = initWithVastVerifications();
+
+                            sinon.stub(deviceInfo, 'getScreenWidth').resolves(1280);
+                            sinon.stub(deviceInfo, 'getScreenHeight').resolves(768);
                             sandbox.stub(om, 'loaded');
                             sandbox.stub(om, 'geometryChange');
                             sandbox.stub(om, 'impression');
                             sandbox.stub(om.getOmidBridge(), 'sendQueuedEvents');
+
                             clock = sinon.useFakeTimers();
                         });
 
-                        it('should call session begin ad events', () => {
-                            om.onEventProcessed('sessionStart');
-
-                            sinon.assert.called(<sinon.SinonSpy>om.loaded);
-                            sinon.assert.notCalled(<sinon.SinonSpy>om.geometryChange);
+                        it('should call session start ad events', () => {
+                            return om.onEventProcessed('sessionStart').then(() => {
+                                sinon.assert.called(<sinon.SinonSpy>om.loaded);
+                                sinon.assert.called(<sinon.SinonSpy>om.impression);
+                                sinon.assert.notCalled(<sinon.SinonSpy>om.geometryChange);
+                            });
                         });
 
                         it('should call session begin ad events for IAS', () => {
-                            om.onEventProcessed('sessionStart', 'IAS');
-
-                            clock.tick(2000);
-                            clock.restore();
-                            sinon.assert.called(<sinon.SinonSpy>om.getOmidBridge().sendQueuedEvents);
-                            sinon.assert.called(<sinon.SinonSpy>om.loaded);
-                            sinon.assert.called(<sinon.SinonSpy>om.geometryChange);
+                            return om.onEventProcessed('sessionStart', 'IAS').then(() => {
+                                clock.tick(1000);
+                                clock.restore();
+                                sinon.assert.called(<sinon.SinonSpy>om.getOmidBridge().sendQueuedEvents);
+                                sinon.assert.called(<sinon.SinonSpy>om.impression);
+                                sinon.assert.called(<sinon.SinonSpy>om.loaded);
+                                sinon.assert.calledWith(<sinon.SinonSpy>om.geometryChange, { height: 768, width: 1280 }, {
+                                    containerGeometry: { height: 768, width: 1280, x: 0, y: 0 },
+                                    geometry: { height: 768, width: 1280, x: 0, y: 0 },
+                                    measuringElement: true,
+                                    onScreenContainerGeometry: { height: 768, obstructions: [], width: 1280, x: 0, y: 0 },
+                                    onScreenGeometry: { height: 768, obstructions: [], width: 1280, x: 0, y: 0 },
+                                    percentageInView: 100,
+                                    reasons: []
+                                });
+                            });
                         });
                     });
 
@@ -197,282 +212,6 @@ import { VastVerificationResource } from 'VAST/Models/VastVerificationResource';
                             sinon.assert.called(<sinon.SinonSpy>om.removeFromViewHieararchy);
                         });
                     });
-                });
-            });
-        });
-
-        describe('Calculating Views Tests', () => {
-
-            om = initWithVastVerifications();
-
-            describe('Calculating PercentageInView', () => {
-                const percentageInViewTests: {
-                    obstruction: IRectangle;
-                    videoView: IRectangle;
-                    screenRectangle: IRectangle;
-                    expectedPercentage: number;
-                    testingCase: string;
-                }[] = [{
-                    obstruction: om.createRectangle(0, 0, 517, 367),
-                    videoView: om.createRectangle(0, 0, 1080, 1800),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 90.23976337448559,
-                    testingCase: 'should calculate video view percentage if video is full screen portrait and obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 517, 367),
-                    videoView: om.createRectangle(0, 0, 1400, 400),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 43.26089285714287,
-                    testingCase: 'should calculate video view percentage if video width exceeds screen width and is obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 517, 367),
-                    videoView: om.createRectangle(0, 0, 800, 2000),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 78.1413125,
-                    testingCase: 'should calculate video view percentage with view obstructed and video height larger than viewport'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 800, 367),
-                    videoView: om.createRectangle(700, 0, 800, 400),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 36.03125,
-                    testingCase: 'should calculate video view percentage if video dimensions are smaller than screen dimensions with x offset and obstruction'
-                },
-                {
-                    obstruction: om.createRectangle(0, 1400, 800, 367),
-                    videoView: om.createRectangle(0, 1500, 800, 400),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 8.25,
-                    testingCase: 'should calculate video view percentage if video dimensions are smaller than screen dimensions with y offset and obstruction'
-                },
-                {
-                    obstruction: om.createRectangle(50, 50, 1280, 3000),
-                    videoView: om.createRectangle(0, 0, 1400, 1800),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 5.615079365079367,
-                    testingCase: 'should calculate video view percentage if video dimensions are larger than screen dimensions and obstruction is larger than screen'
-                },
-                {
-                    obstruction: om.createRectangle(1081, 1801, 1280, 3000),
-                    videoView: om.createRectangle(0, 0, 1400, 1800),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 77.14285714285715,
-                    testingCase: 'should calculate video view percentage if video dimensions are larger than screen dimensions and obstruction is not in screen view'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 0, 0),
-                    videoView: om.createRectangle(0, 0, 1400, 1800),
-                    screenRectangle: om.createRectangle(0, 0, 1080, 1800),
-                    expectedPercentage: 77.14285714285715,
-                    testingCase: 'should calculate video view percentage if video dimensions are larger than screen dimensions and obstruction is ZeroRect'
-                }];
-                percentageInViewTests.forEach((t) => {
-                    it(t.testingCase, () => {
-                        const calculatedPercentage = om.calculatePercentageInView(t.videoView, t.obstruction, t.screenRectangle);
-                        assert.equal(calculatedPercentage, t.expectedPercentage);
-                    });
-                });
-            });
-
-            describe('Calculating ObstructionOverlapPercentage', () => {
-
-                const obstructionOverlapTests: {
-                    obstruction: IRectangle;
-                    videoView: IRectangle;
-                    expectedPercentage: number;
-                    testingCase: string;
-                }[] = [{
-                    obstruction: om.createRectangle(0, 0, 517, 367),
-                    videoView: om.createRectangle(0, 0, 1280, 768),
-                    expectedPercentage: 19.301249186197918,
-                    testingCase: 'should calculate video view percentage if full screen landscape video is obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 517, 367),
-                    videoView: om.createRectangle(0, 0, 768, 1280),
-                    expectedPercentage: 19.301249186197918, //Fix this
-                    testingCase: 'should calculate video view percentage if full screen portrait video is obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(0, 350, 768, 500),
-                    videoView: om.createRectangle(0, 350, 768, 500),
-                    expectedPercentage: 100,
-                    testingCase: 'should calculate video view percentage if sized-to-fit landscape video is fully obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(20, 20, 730, 900),
-                    videoView: om.createRectangle(0, 350, 768, 500),
-                    expectedPercentage: 95.05208333333334,
-                    testingCase: 'should calculate video view percentage if sized-to-fit portrait video is mostly obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(20, 20, 517, 367),
-                    videoView: om.createRectangle(0, 350, 768, 500),
-                    expectedPercentage: 4.981510416666667,
-                    testingCase: 'should calculate video view percentage if sized-to-fit portrait video is obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 517, 367),
-                    videoView: om.createRectangle(0, 350, 768, 500),
-                    expectedPercentage: 2.2888020833333336,
-                    testingCase: 'should calculate video view percentage if sized-to-fit portrait video is partially obstructed'
-                },
-                {
-                    obstruction: om.createRectangle(0, 0, 517, 349),
-                    videoView: om.createRectangle(0, 350, 768, 500),
-                    expectedPercentage: 0,
-                    testingCase: 'should calculate overlap percentage as 0 if obstruction does not pass over video view'
-                }];
-
-                obstructionOverlapTests.forEach((t) => {
-                    it(t.testingCase, () => {
-                        const calculatedPercentage = om.calculateObstructionOverlapPercentage(t.videoView, t.obstruction);
-                        assert.equal(calculatedPercentage, t.expectedPercentage);
-                    });
-                });
-            });
-
-            describe('Calculating Vast AdView', () => {
-                it('should return the adview in landscape', () => {
-                    const calculatedAdView: IAdView = om.calculateVastAdView(100, [], 200, 100, false, []);
-                    const testAdView: IAdView = {
-                        percentageInView: 100,
-                        geometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100
-                        },
-                        onScreenGeometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100,
-                            obstructions: []
-                        },
-                        measuringElement: false,
-                        reasons: []
-                    };
-
-                    assert.equal(JSON.stringify(calculatedAdView), JSON.stringify(testAdView));
-                });
-
-                it('should return the adview in portrait', () => {
-                    const calculatedAdView: IAdView = om.calculateVastAdView(100, [], 100, 200, false, []);
-                    const testAdView: IAdView = {
-                        percentageInView: 100,
-                        geometry: {
-                            x: 0,
-                            y: 0,
-                            width: 100,
-                            height: 200
-                        },
-                        onScreenGeometry: {
-                            x: 0,
-                            y: 0,
-                            width: 100,
-                            height: 200,
-                            obstructions: []
-                        },
-                        measuringElement: false,
-                        reasons: []
-                    };
-
-                    assert.equal(JSON.stringify(calculatedAdView), JSON.stringify(testAdView));
-                });
-
-                it('should return new adview with an obstruction', () => {
-                    const obstructionRectangle = {
-                        x: 0,
-                        y: 0,
-                        width: 50,
-                        height: 50
-                    };
-
-                    const calculatedAdView: IAdView = om.calculateVastAdView(50, [ObstructionReasons.OBSTRUCTED], 200, 100, false, [obstructionRectangle]);
-                    const testAdView: IAdView = {
-                        percentageInView: 50,
-                        geometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100
-                        },
-                        onScreenGeometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100,
-                            obstructions: [obstructionRectangle]
-                        },
-                        measuringElement: false,
-                        reasons: [ObstructionReasons.OBSTRUCTED]
-                    };
-                    assert.equal(JSON.stringify(calculatedAdView), JSON.stringify(testAdView));
-                });
-
-                it('should return the adview with measuringElementAvailable', () => {
-                    const calculatedAdView: IAdView = om.calculateVastAdView(100, [], 200, 100, true, []);
-                    const testAdView: IAdView = {
-                        percentageInView: 100,
-                        geometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100
-                        },
-                        onScreenGeometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100,
-                            obstructions: []
-                        },
-                        measuringElement: true,
-                        reasons: [],
-                        containerGeometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100
-                        },
-                        onScreenContainerGeometry: {
-                            x: 0,
-                            y: 0,
-                            width: 200,
-                            height: 100,
-                            obstructions: []
-                        }
-                    };
-                    assert.equal(JSON.stringify(calculatedAdView), JSON.stringify(testAdView));
-                });
-
-                it('should return the adview based on videoViewRectangle', () => {
-                    om.setVideoViewRectangle(om.createRectangle(0, 200, 300, 300));
-                    const calculatedAdView: IAdView = om.calculateVastAdView(100, [], 200, 100, false, []);
-
-                    const testAdView: IAdView = {
-                        percentageInView: 100,
-                        geometry: {
-                            x: 0,
-                            y: 200,
-                            width: 300,
-                            height: 300
-                        },
-                        onScreenGeometry: {
-                            x: 0,
-                            y: 200,
-                            width: 300,
-                            height: 300,
-                            obstructions: []
-                        },
-                        measuringElement: false,
-                        reasons: []
-                    };
-
-                    assert.equal(JSON.stringify(calculatedAdView), JSON.stringify(testAdView));
                 });
             });
         });